+++ database/migrations/2025_07_01_105218_add_image_to_announcements_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('announcements', function (Blueprint $table) {
-            $table->dropColumn('image');
-        });
+        if (Schema::hasColumn('announcements', 'image')) {
+            Schema::table('announcements', function (Blueprint $table) {
+                $table->dropColumn('image');
+            });
+        }
     }
 };
